# ARROW-10763: [Rust] Speed up take for primitive / boolean for non-null arrays