--- a/rust/arrow/src/compute/kernels/take.rs
+++ b/rust/arrow/src/compute/kernels/take.rs
@@ -220,7 +220,7 @@ where
     let nulls;
 
     if null_count == 0 {
-        // Just taking indices without null checking
+        // Take indices without null checking
         for i in 0..data_len {
             let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
                 ArrowError::ComputeError("Cast to usize failed".to_string())
@@ -284,36 +284,62 @@ where
     let array = values.as_any().downcast_ref::<BooleanArray>().unwrap();
 
     let num_byte = bit_util::ceil(data_len, 8);
-    let mut null_buf = MutableBuffer::new(num_byte).with_bitset(num_byte, true);
     let mut val_buf = MutableBuffer::new(num_byte).with_bitset(num_byte, false);
 
-    let null_slice = null_buf.data_mut();
     let val_slice = val_buf.data_mut();
 
-    (0..data_len).try_for_each::<_, Result<()>>(|i| {
-        let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
-            ArrowError::ComputeError("Cast to usize failed".to_string())
+    let null_count = array.null_count();
+
+    let nulls;
+    if null_count == 0 {
+        (0..data_len).try_for_each::<_, Result<()>>(|i| {
+            let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
+                ArrowError::ComputeError("Cast to usize failed".to_string())
+            })?;
+
+            if array.value(index) {
+                bit_util::set_bit(val_slice, i);
+            }
+
+            Ok(())
         })?;
 
-        if array.is_null(index) {
-            bit_util::unset_bit(null_slice, i);
-        } else if array.value(index) {
-            bit_util::set_bit(val_slice, i);
-        }
+        nulls = indices.data_ref().null_buffer().cloned();
+    } else {
+        let mut null_buf = MutableBuffer::new(num_byte).with_bitset(num_byte, true);
+        let null_slice = null_buf.data_mut();
 
-        Ok(())
-    })?;
+        (0..data_len).try_for_each::<_, Result<()>>(|i| {
+            let index = ToPrimitive::to_usize(&indices.value(i)).ok_or_else(|| {
+                ArrowError::ComputeError("Cast to usize failed".to_string())
+            })?;
 
-    let nulls = match indices.data_ref().null_buffer() {
-        Some(buffer) => buffer_bin_and(buffer, 0, &null_buf.freeze(), 0, indices.len()),
-        None => null_buf.freeze(),
-    };
+            if array.is_null(index) {
+                bit_util::unset_bit(null_slice, i);
+            } else if array.value(index) {
+                bit_util::set_bit(val_slice, i);
+            }
+
+            Ok(())
+        })?;
+
+        nulls = match indices.data_ref().null_buffer() {
+            Some(buffer) => Some(buffer_bin_and(
+                buffer,
+                0,
+                &null_buf.freeze(),
+                0,
+                indices.len(),
+            )),
+            None => Some(null_buf.freeze()),
+        };
+    }
 
     let data = ArrayData::new(
         DataType::Boolean,
         indices.len(),
         None,
-        Some(nulls),
+        nulls,
         0,
         vec![val_buf.freeze()],
         vec![],